requirements.txt
pandas
plotly
streamlit
langchain
transformers
python-dotenv
langchain_google_genai
tensorflow
tf-keras
opencv-python
openai-whisper
kafka-python
pyspark
scikit-learn
faiss-cpu
matplotlib
statsmodels
torch
efficientnet_pytorch
facenet_pytorch
whisper
torchvision
torchtext
langchain_community
sentence-transformers
pyvis
prometheus-client 
pynvml
psutil
moviepy
pymongo
pytube
pafy
yt-dlp
m3u8
websockets
streamlink
ffmpeg-python
youtube-dl
geopandas
folium
geopy
streamlit-folium
googletrans
httpx
deep-translator
langdetect

numpy
facenet-pytorch
aiokafka
prometheus_client
pymongo
python-dotenv
networkx
google-generativeai